Move eigenvalue helpers to leading_eigenv_approx module
Move get_largest_eigenvalue and approximate_largest_eigenvalue from
versatility.py into leading_eigenv_approx.py where they belong.
versatility.py now imports them from there.
