+++ src/MuxVizPy/versatility.py
@@ -11,101 +11,9 @@ import logging
 from typing import Optional
 
 from MuxVizPy import leading_eigenv_approx
+from MuxVizPy.leading_eigenv_approx import get_largest_eigenvalue, approximate_largest_eigenvalue
 from MuxVizPy import build
 from MuxVizPy.utils import parsing as parsing_utils
-
-
-# ---------------------------------------------------------------------------
-# Eigenvalue helpers (integrated from hornet/node_based)
-# ---------------------------------------------------------------------------
-
-def get_largest_eigenvalue(adj: sps.spmatrix, logger: Optional[logging.Logger] = None) -> tuple[float, np.ndarray]:
-    """
-    Compute the largest eigenvalue and corresponding eigenvector of a sparse matrix.
-
-    Parameters
-    ----------
-    adj : scipy.sparse matrix
-        Adjacency matrix.
-    logger : logging.Logger, optional
-        Logger for debug information.
-
-    Returns
-    -------
-    tuple
-        (largest eigenvalue, corresponding eigenvector)
-    """
-    eigvals, eigvecs = sps.linalg.eigs(adj, k=1, which="LM", return_eigenvectors=True)
-    real_mask = np.abs(np.imag(eigvals)) < 1e-6
-
-    if not np.any(real_mask):
-        if logger:
-            logger.warning("Warning! Complex numbers in the leading eigenvalue.")
-        lam = float(np.real(eigvals[0]))
-        vec = np.real(eigvecs[:, 0])
-    else:
-        lam = float(np.real(eigvals[0]))
-        vec = np.real(eigvecs[:, 0])
-
-    if not np.allclose(np.imag(vec), 0):
-        if logger:
-            logger.warning("Warning! Complex numbers in the leading eigenvector.")
-
-    vec[(vec > -1e-12) & (vec < 1e-12)] = 0.0
-
-    non_zero_vec = vec[vec != 0]
-    if len(non_zero_vec) > 0 and np.all(non_zero_vec < 0):
-        vec = -vec
-
-    if logger and logger.isEnabledFor(logging.DEBUG):
-        logger.debug(f"Largest eigenvalue: {lam}, eigenvector min: {vec.min()}, max: {vec.max()}, mean: {vec.mean()}")
-
-    return lam, vec
-
-
-def approximate_largest_eigenvalue(
-    adj: sps.spmatrix, maxiter: int = 1000, tol: float = 1e-6,
-    cval: Optional[float] = None, alpha: float = 0.85,
-) -> tuple[float, np.ndarray]:
-    """
-    Approximate the largest eigenvalue using the power method with PageRank-style shift.
-
-    Parameters
-    ----------
-    adj : scipy.sparse matrix
-        Adjacency matrix.
-    maxiter : int
-        Maximum iterations.
-    tol : float
-        Convergence tolerance.
-    cval : float or None
-        Constant shift value (default: (1 - alpha) / n).
-    alpha : float
-        Damping factor.
-
-    Returns
-    -------
-    tuple
-        (approximate eigenvalue, eigenvector)
-    """
-    n = adj.shape[0]
-    x = np.random.rand(n)
-    if cval is None:
-        cval = (1 - alpha) / n
-
-    for counter in range(maxiter):
-        y = alpha * (adj @ x) + cval * np.ones_like(x) * x.sum()
-        x_new = y / y.sum()
-        delta = np.linalg.norm(x_new - x)
-        if delta < tol:
-            x = x_new
-            break
-        x = x_new
-
-    num = alpha * (x @ (adj @ x)) + cval * (x.sum() ** 2)
-    den = x @ x
-    lam = num / den
-    return lam, x
 
 
 # ---------------------------------------------------------------------------
